Split weapons and spells rating code from combat actions
