+++ apps/openmw/CMakeLists.txt
@@ -84,7 +84,7 @@ add_openmw_dir (mwmechanics
     drawstate spells activespells npcstats aipackage aisequence aipursue alchemy aiwander aitravel aifollow aiavoiddoor aibreathe
     aiescort aiactivate aicombat repair enchanting pathfinding pathgrid security spellsuccess spellcasting
     disease pickpocket levelledlist combat steering obstacle autocalcspell difficultyscaling aicombataction actor summoning
-    character actors objects aistate coordinateconverter trading aiface
+    character actors objects aistate coordinateconverter trading aiface weaponpriority spellpriority
     )
 
 add_openmw_dir (mwstate
